# Make `rangeConsent` a struct type